Patch the list command to include the new Platformify command.
We need to override the platform list command on the Go side, to be able to include the new command (platform project:init) in
the list of commands.

This should include both platform list and platform list project.

Technical details

Proxy to the legacy list command using --format=json Parse the JSON response
Add the extra commands, if needed
Return the enhanced result in the correct format (the one the user asked for)